requirements.txt
Flask-Bootstrap4==4.0.2
Flask-Login==0.6.2
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
Flask==2.2.2
SQLAlchemy==1.4.44
Flask-WTF==1.0.1
Flask-Migrate==4.0.0
